+++ R/install.R
@@ -1,4 +1,34 @@
-#' Install TensorFlow and Keras, including all Python dependencies
+#' Install TensorFlow and Legacy Keras, including all Python dependencies
+#'
+#' This function can be used to create a persistent virtual environment for usage
+#' with Legacy Keras. New code is recommended to use the `keras3` package which
+#' works automatically and does any special configuration like this.
+#'
+#' Note that recent versions of reticulate will automatically manage
+#' environments if dependencies are declared with `reticulate::py_require()`.
+#' Instead of using this `install_keras()` function, new users are first
+#' encouraged to use `keras3`. If that's not an option, instead of creating a
+#' persistent venv, you can this code at the start of the R session, before
+#' loading keras:
+#'
+#' ```r
+#' # declare requirements for legacy keras (tf-keras)
+#' reticulate::py_require(c("tensorflow", "tf-keras", "numpy<2"))
+#'
+#' # also declare optional dependencies
+#' reticulate::py_require(c(
+#'   "tensorflow-hub", "tensorflow-datasets", "scipy",
+#'   "requests", "Pillow", "h5py", "pandas", "pydot"
+#' ))
+#' library(keras)
+#' ```
+#'
+#' Or more simply, call `py_require_legacy_keras()`
+#'
+#' ```r
+#' library(keras)
+#' py_require_legacy_keras()
+#' ```
 #'
 #' This function will install Tensorflow and all Keras dependencies. This is a
 #' thin wrapper around [`tensorflow::install_tensorflow()`], with the only
@@ -44,7 +74,7 @@ install_keras <- function(method = c("auto", "virtualenv", "conda"),
     method = method,
     conda = conda,
     version = tensorflow,
-    extra_packages = pkgs,
+    extra_packages = c(pkgs, "tf-keras"),
     # envname = envname,
     # new_env = new_env,
     ...