+++ R/package.R
@@ -94,11 +94,44 @@ use_backend <- function(backend = c("tensorflow", "cntk", "theano", "plaidml")) 
 keras <- NULL
 
 
-.onLoad <- function(libname, pkgname) {
-
+.onAttach <- function(libname, pkgname) {
   packageStartupMessage(
     "The keras package is deprecated. Use the keras3 package instead."
   )
+}
+
+#' @export
+#' @rdname install_keras
+py_require_legacy_keras <- function(extra_packages = TRUE) {
+  Sys.setenv(TF_USE_LEGACY_KERAS = 1L)
+  reticulate::py_require(
+    c(
+      "tf-keras", "tensorflow", "numpy<2",
+      if (extra_packages) {
+        c(
+          "tensorflow-hub",
+          "tensorflow-datasets",
+          "scipy",
+          "requests",
+          "Pillow",
+          "h5py",
+          "pandas",
+          "pydot"
+        )
+      }
+    )
+  )
+}
+
+.onLoad <- function(libname, pkgname) {
+
+  Sys.setenv(TF_USE_LEGACY_KERAS = 1L)
+
+  ## In a future update, we will declare these dependencies by default.
+  ## We're making this opt-in for now to minimize breaking user code
+  ## and give package maintainers an opportunity to make updates
+  if (tolower(Sys.getenv("PY_REQUIRE_LEGACY_KERAS")) %in% c("1", "true", "yes"))
+    py_require_legacy_keras()
 
   # resolve the implementation module (might be keras proper or might be tensorflow)
   implementation_module <- resolve_implementation_module()
@@ -133,6 +166,10 @@ keras <- NULL
     },
 
     on_error = function(e) {
+      emit <- base::message
+      emit(
+        "To use legacy Keras via py_require(), call py_require_legacy_keras() at the start of the R session."
+      )
       if (is_tensorflow_implementation())
         stop(tf_config()$error_message, call. = FALSE)
       else {
@@ -155,6 +192,9 @@ keras <- NULL
 
     # replace "tensorflow.python.keras.*" with "keras.*"
     classes <- sub(paste0("^", module), "keras", classes)
+
+    # replace legacy keras module name: tf_keras -> keras
+    classes <- sub("^tf_keras\\.", "keras.", classes)
 
     # All python symbols moved in v2.13 under .src
     # Preserve the original symbols for compatability with keras3,